distro/fedora: move imageType implementation to separate file
